requirements.txt
# Django Framework
Django==4.2.7

# Database Drivers
# For MySQL (bonus requirement - uncomment when needed)
mysqlclient==2.2.0

# For PostgreSQL (uncomment when needed)
# psycopg2==2.9.9
# psycopg2-binary==2.9.9

# Forms and UI
django-crispy-forms==2.0
crispy-bootstrap5==0.7

# Development Tools
django-debug-toolbar==4.2.0

# Production Dependencies (uncomment for deployment)
# gunicorn==21.2.0
# whitenoise==6.6.0

# Environment Management
python-decouple==3.8

# Testing
coverage==7.3.2

# Utils
Pillow==10.1.0
python-dateutil==2.8.2

# Static Files (for production)
# django-storages==1.14.2
# boto3==1.34.0  # For AWS S3